+++ CHANGELOG.markdown
@@ -7,9 +7,12 @@
 * Add `c => Boring (Dict c)` instance
 * Remove `Lifting` instances for `ErrorT` and `ListT`, which were removed
   in `transformers-0.6.*`.
+* Add the `Data.Constraint.Char` module, which contains utilities for working
+  with `KnownChar` constraints. This module is only available on GHC 9.2 or
+  later.
 * Add `unsafeAxiom` to `Data.Constraint.Unsafe`.
-* Add `unsafeSNat` and `unsafeSSymbol` to `Data.Constraint.Unsafe` when building
-  with `base-4.18` (GHC 9.6) or later.
+* Add `unsafeSChar`, `unsafeSNat`, and `unsafeSSymbol` to
+  `Data.Constraint.Unsafe` when building with `base-4.18` (GHC 9.6) or later.
 
 0.13.4 [2022.05.19]
 -------------------